backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
minio
tqdm
websockets
pyfiglet

